+++ lib/arachni/rpc/server/framework.rb
@@ -145,14 +145,14 @@ class Framework < ::Arachni::Framework
     # You don't need to call this if you've let the scan complete.
     #
     def clean_up( &block )
-        if @cleaned_up
+        if @rpc_cleaned_up
             # Don't shutdown the BrowserCluster here, its termination will be
             # handled by Instance#shutdown.
             block.call false if block_given?
             return false
         end
 
-        @cleaned_up       = true
+        @rpc_cleaned_up   = true
         @extended_running = false
 
         r = super( false )